requirements.txt
appdirs==1.4.0
astroid==1.4.9
colorama==0.3.7
Django==1.10.5
django-cors-headers==2.0.1
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.5.3
elasticsearch==5.2.0
elasticsearch-dsl==5.1.0
isort==4.2.5
lazy-object-proxy==1.2.2
mccabe==0.6.1
mongoengine==0.11.0
nltk==3.2.2
packaging==16.8
pylint==1.6.5
pylint-django==0.7.2
pylint-plugin-utils==0.2.4
pymongo==3.4.0
pyparsing==2.1.10
python-dateutil==2.6.0
six==1.10.0
urllib3==1.20
wrapt==1.10.8
